+++ src/controller/src/controller/transaction_controller.hpp
@@ -10,6 +10,7 @@ namespace store
     class ITransactionStore;   // Forward declaration
     class IAccountStore;       // Forward declaration
     class IStockStore;         // Forward declaration
+    class IOptionStore;        // Forward declaration
 }   // namespace store
 
 namespace cmd
@@ -45,6 +46,8 @@ namespace controller
         std::shared_ptr<store::IAccountStore> _accountStore;
         /// Reference to the stock store
         std::shared_ptr<store::IStockStore> _stockStore;
+        /// Reference to the option store
+        std::shared_ptr<store::IOptionStore> _optionStore;
 
         struct UIElements;
         /// A unique pointer to the UI elements used by the controller,
@@ -59,6 +62,7 @@ namespace controller
             const std::shared_ptr<store::ITransactionStore>& transactionStore,
             const std::shared_ptr<store::IAccountStore>&     accountStore,
             const std::shared_ptr<store::IStockStore>&       stockStore,
+            const std::shared_ptr<store::IOptionStore>&      optionStore,
             QStackedWidget*                                  stackedWidget
         );
         ~TransactionController() override;